requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anthropic==0.31.0
anyio==4.4.0
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
eval_type_backport==0.2.0
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
groq==0.9.0
h11==0.14.0
http-exceptions==0.2.10
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
Jinja2==3.1.4
jiter==0.5.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
numpy==2.0.0
openai==1.35.13
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pip==24.0
protobuf==5.27.2
pyarrow==16.1.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
streamlit==1.37.1
tabulate==0.9.0
tenacity==8.5.0
together==1.2.1
tokenizers==0.19.1
toml==0.10.2
toolhouse==1.3.0
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
yarl==1.9.4
arize-phoenix-otel
openinference-instrumentation-groq
openinference-instrumentation-openai